+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Castle.Facilities.NHibernateIntegration - Changelog
 
 
-## 5.3.0 (2022-07-xx)
+## 5.3.0 (2022-08-xx)
+
+Improvements:
+- Added **`appsettings.json`** (```Microsoft.Extensions.Configuration```) support.
 
 Breaking Changes:
 - Upgraded to **`.NET 6.0`** and **`.NET Framework 4.8`**.